[HttpKernel] Moved classes to existing namespace
