+++ src/Symfony/Component/HttpKernel/Tests/Controller/ArgumentResolverTest.php
@@ -19,8 +19,8 @@ use Symfony\Component\HttpFoundation\Session\Storage\MockArraySessionStorage;
 use Symfony\Component\HttpKernel\Controller\ArgumentResolver;
 use Symfony\Component\HttpKernel\Controller\ArgumentResolver\RequestAttributeValueResolver;
 use Symfony\Component\HttpKernel\Controller\ArgumentValueResolverInterface;
-use Symfony\Component\HttpKernel\ControllerConfiguration\Configuration\QueryParam;
-use Symfony\Component\HttpKernel\ControllerConfiguration\ConfigurationList;
+use Symfony\Component\HttpKernel\Controller\Configuration\QueryParam;
+use Symfony\Component\HttpKernel\Controller\ConfigurationList;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadataFactory;
 use Symfony\Component\HttpKernel\Tests\Fixtures\Controller\AnnotatedController;
 use Symfony\Component\HttpKernel\Tests\Fixtures\Controller\ExtendingRequest;
